refactor(sidebar): extract SidebarTabDefinition and TabPopoutMixin to fix file-size budget
